CV Roasting Flask/requirements.txt
Flask
langchain-ollama
langchain-core
python-docx